reversed changes from float back to int

+++ src/Helper/ElasticExportHelper.hh
@@ -503,11 +503,11 @@ class ElasticExportHelper
 	 * Get category branch marketplace for a custom branch id.
 	 * @param  int $categoryhId
 	 * @param  int $plentyId
-	 * @param  float $marketplaceId
+	 * @param  int $marketplaceId
 	 * @param  string $marketplaceSubId
 	 * @return string
 	 */
-	public function getCategoryMarketplace(int $categoryhId, int $plentyId, float $marketplaceId, float $marketplaceSubId = 0.0):string
+	public function getCategoryMarketplace(int $categoryhId, int $plentyId, int $marketplaceId, float $marketplaceSubId = 0.0):string
 	{
 		$categoryBranchMarketplace = $this->categoryBranchMarketplaceRepository->findCategoryBranchMarketplace($categoryhId, $plentyId, $marketplaceId, $marketplaceSubId);
